chore: remove duplicate Backend folder, consolidate into backend
