--- Backend/package.json
@@ -1,19 +1,0 @@
-{
-  "dependencies": {
-    "dotenv": "^16.0.0",
-    "express": "^4.18.0",
-    "pg": "^8.11.0"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.0",
-    "@types/node": "^20.0.0",
-    "@types/pg": "^8.10.0",
-    "ts-node": "^10.9.0",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.0.0"
-  },
-  "scripts": {
-  "migrate": "ts-node src/db/migrate.ts"
-}
-}
-
