--- Backend/src/db/client.ts
@@ -1,17 +1,0 @@
-import { Pool } from 'pg';
-import dotenv from 'dotenv';
-
-dotenv.config();
-
-const pool = new Pool({
-  host:     process.env.DB_HOST,
-  port:     Number(process.env.DB_PORT),
-  database: process.env.DB_NAME,
-  user:     process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-});
-
-pool.on('connect', () => console.log('✅ PostgreSQL connected'));
-pool.on('error',   (err) => console.error('❌ DB error:', err));
-
-export default pool;
