+++ backend/.env.example
@@ -1,1 +1,5 @@
-// Environment variables example file
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=fraud_detection
+DB_USER=postgres
+DB_PASSWORD=your_password
